+++ previous-versions/files/src/codatfiles/files.py
@@ -4,7 +4,7 @@ from .sdkconfiguration import SDKConfiguration
 from codatfiles import utils
 from codatfiles.models import errors, operations, shared
 from enum import Enum
-from typing import Any, Optional
+from typing import Optional
 
 class DownloadFilesAcceptEnum(str, Enum):
     APPLICATION_JSON = "application/json"
@@ -111,7 +111,7 @@ class Files:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[list[dict[str, Any]]])
+                out = utils.unmarshal_json(http_res.text, Optional[list[shared.File]])
                 res.files = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)