+++ previous-versions/files/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] previous-versions/files
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-files/0.1.0 - previous-versions/files
+
+## 2023-10-16 01:24:38
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.1] previous-versions/files
+### Releases
+- [PyPI v0.1.1] https://pypi.org/project/codat-files/0.1.1 - previous-versions/files